Added model data and configurations. All good to go now, just need to record video.

+++ TobiiInputInterface/TobiiInputInterface/TobiiToOSCRunner.cs
@@ -72,7 +72,7 @@ namespace TobiiInputInterface
 
         static OscMessage buildMessageFromPoint(Point p)
         {
-            OscMessage message = new SharpOSC.OscMessage("/wek/inputs", p.x, p.y);
+            OscMessage message = new SharpOSC.OscMessage("/wek/inputs", (float)p.x, (float)p.y);
             return message;
         }
